requirements.txt
streamlit
google-generativeai
pymupdf
pandas
openpyxl